Add failing test case
Sorry for the PHPUnit, it's the best way I know :'(

Also, would be nice to be able to define factories in the test itself,
unfortunate to have the definitions live in a separate file outside the
test case. Maybe consider exposing a method on the `Factory` class to
define factories manually?